+++ packages/core/src/components/legend/legend.interfaces.ts
@@ -1,0 +1,14 @@
+export interface LegendTabItem {
+  label: string;
+  id: string;
+  active: boolean;
+}
+
+export interface LegendContentSwitchEvent {
+  originalEvent: MouseEvent | KeyboardEvent;
+  isModifiedEvent: boolean;
+}
+
+export interface LegendCloseEvent {
+  originalEvent?: MouseEvent | Event;
+}