+++ packages/dso-toolkit/src/components/legend/index.ts
@@ -1,0 +1,3 @@
+export * from "./legend.args";
+export * from "./legend.models";
+export * from "./legend.stories-of";