+++ packages/dso-toolkit/src/components/legend/legend.models.ts
@@ -1,0 +1,22 @@
+export interface Legend<TemplateFnReturnType> {
+  content: TemplateFnReturnType;
+  tabItems: LegendTabItem[];
+  dsoContentSwitch?: (e: CustomEvent<LegendContentSwitchEvent>) => void;
+  dsoClose?: (e: CustomEvent<LegendCloseEvent>) => void;
+}
+
+export interface LegendTabItem {
+  label: string;
+  id: string;
+  active: boolean;
+}
+
+export interface LegendContentSwitchEvent {
+  originalEvent: MouseEvent | KeyboardEvent;
+  isModifiedEvent: boolean;
+  tabItem: LegendTabItem;
+}
+
+export interface LegendCloseEvent {
+  originalEvent?: MouseEvent | Event;
+}